pyproject.toml
[project]
name = "abca4-campaign"
version = "0.1.0"
description = "ABCA4 variant triage campaign (notebooks + pipeline)"
authors = [{ name = "ABCA4 Team" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26,<2.0",
  "scipy>=1.12",
  "requests>=2.31.0",
  "pyyaml>=6.0",
  "biopython>=1.81",
  "pysam>=0.23.3",
  "pyensembl>=2.2",
  "mlflow>=2.14",
  "marimo[recommended]>=0.17.8",
  "plotly>=6.4.0",
  "jinja2>=3.1.0",
  "invoke>=2.2.0",
  "groq>=0.4.1",
  "playwright>=1.56.0",
]

[project.optional-dependencies]
interactive = ["marimo>=0.17.8", "plotly>=5.15.0"]

[tool.uv]

[dependency-groups]
dev = ["pytest>=7"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
